feat: create build workflow for sub-project
